+++ src/Behat/Behat/Definition/Printer/ConsoleDefinitionPrinter.php
@@ -13,6 +13,7 @@ namespace Behat\Behat\Definition\Printer;
 use Behat\Behat\Definition\Definition;
 use Behat\Behat\Definition\Pattern\PatternTransformer;
 use Behat\Behat\Definition\Translator\DefinitionTranslator;
+use Behat\Gherkin\Keywords\KeywordsInterface;
 use Behat\Testwork\Suite\Suite;
 use Symfony\Component\Console\Formatter\OutputFormatterStyle;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -36,20 +37,27 @@ abstract class ConsoleDefinitionPrinter implements DefinitionPrinter
      * @var DefinitionTranslator
      */
     private $translator;
+    /**
+     * @var KeywordsInterface
+     */
+    private $keywords;
 
     /**
      * Initializes printer.
      *
-     * @param OutputInterface      $output
-     * @param PatternTransformer   $patternTransformer
+     * @param OutputInterface     $output
+     * @param PatternTransformer  $patternTransformer
      * @param DefinitionTranslator $translator
+     * @param KeywordsInterface   $keywords
      */
     public function __construct(
         OutputInterface $output,
         PatternTransformer $patternTransformer,
-        DefinitionTranslator $translator
+        DefinitionTranslator $translator,
+        KeywordsInterface $keywords
     ) {
         $this->output = $output;
+        $this->keywords = $keywords;
         $this->patternTransformer = $patternTransformer;
         $this->translator = $translator;
 
@@ -73,6 +81,15 @@ abstract class ConsoleDefinitionPrinter implements DefinitionPrinter
     {
         $this->output->writeln($text);
         $this->output->writeln('');
+    }
+
+    final protected function getDefinitionType(Definition $definition)
+    {
+        $this->keywords->setLanguage($this->translator->getLocale());
+
+        $method = 'get'.ucfirst($definition->getType()).'Keywords';
+
+        return $this->keywords->$method();
     }
 
     /**